comment upgrade

+++ 20171221/post/app/controllers/post_controller.rb
@@ -1,7 +1,7 @@
 class PostController < ApplicationController
   before_action :authorize, except: [:index]
   def index
-    @post=PostList.all
+    @posts=PostList.all
   end
 
   def edit
@@ -21,7 +21,7 @@ class PostController < ApplicationController
 
   def update
     post=PostList.find(params[:id])
-    post.update(title: params[:title],user_id: current_user.id, text: params[:text])
+    post.update(title: params[:title],user_id: post.user_id , text: params[:text])
     redirect_to "/post/show/#{params[:id]}"
   end
 
@@ -37,18 +37,33 @@ class PostController < ApplicationController
     @p=PostList.find(params[:id])
     @c=Comment.all
   end
+
   def add_comment
-    Comment.create(name: params[:name], comm: params[:text], post_list_id: params[:id] )
+    Comment.create(name: params[:name], comm: params[:text], post_list_id: params[:id],user_id: current_user.id )
     redirect_to :back
   end
   def comdel
     comment = Comment.find(params[:id])
     post = comment.post_list_id
-    comment.destroy
+    if comment.user_id==current_user.id
+      comment.destroy
+    end
     redirect_to "/post/show/#{post}"
   end
   def comedit
-    @comm=Comment.find(params[:id]) 
+    @comm=Comment.find(params[:id])
+    if @comm.user_id == current_user.id
+
+    else
+      redirect_to "/post/show/#{@comm.post_list_id}"
+    end
   end
+  def comupdate
+    comm=Comment.find(params[:id])
+    comm.update(name: params[:name], comm: params[:text], post_list_id: comm.post_list_id)
+
+    redirect_to "/post/show/#{comm.post_list_id}"
+  end
+
 
 end
